+++ resources/views/dashboard/index.php
@@ -1,39 +1,45 @@
 <?php
+
 declare(strict_types=1);
 
 /**
- * @see DashboardController::actionIndex()
+ * @see \Hirtz\Skeleton\Modules\Admin\Controllers\DashboardController::actionIndex()
  * @var View $this
- * @var array $panels
- * @var AnakinAssetBundle $bundle
  */
 
 use Hirtz\Anakin\Assets\AnakinAssetBundle;
-use Hirtz\Skeleton\Modules\Admin\Controllers\DashboardController;
+use Hirtz\Skeleton\Modules\Admin\Widgets\DirectoryAlert;
+use Hirtz\Skeleton\Modules\Admin\Widgets\EnvironmentAlert;
+use Hirtz\Skeleton\Modules\Admin\Widgets\MigrationAlert;
 use Hirtz\Skeleton\Modules\Admin\Widgets\Navs\DashboardHeader;
 use Hirtz\Skeleton\Web\View;
+use Hirtz\Skeleton\Widgets\Navs\Nav;
+use Hirtz\Skeleton\Widgets\Navs\NavItem;
 use Hirtz\Skeleton\Widgets\Panels\Dashboard;
 
+$this->title(Yii::t('skeleton', 'DASHBOARD_NAV_ITEM_LABEL'));
+
+/** @var AnakinAssetBundle $bundle */
 $bundle = $this->getAssetManager()->getBundle(AnakinAssetBundle::class);
 $logoUrl = $bundle->getLogoUrl();
 
 echo DashboardHeader::make();
+echo MigrationAlert::make();
+echo DirectoryAlert::make();
+echo EnvironmentAlert::make();
 echo Dashboard::make();
 ?>
 <div class="text-center">
     <div class="home-wrap">
-        <?php
-        if ($logoUrl) {
-            ?>
+        <?php if ($logoUrl) { ?>
             <div class="home-logo">
                 <img src="<?= $logoUrl; ?>" alt="<?= Yii::$app->name; ?>">
             </div>
-            <?php
-        } ?>
+        <?php } ?>
         <div class="home-welcome">
             <h1>
                 <?= Yii::t('anakin', 'ANAKIN_DASHBOARD_GREETING', [
-                    'name' => Yii::$app->getUser()->getIdentity()->getUsername(),
+                    'name' => Yii::$app->getUser()->getIdentity()?->getUsername(),
                 ]); ?>
             </h1>
             <p>
@@ -41,34 +47,11 @@ echo Dashboard::make();
             </p>
         </div>
     </div>
-    <?php
-    foreach ($panels as $panel) {
-        ?>
-        <?= Nav::widget([
-            'items' => $panel['items'],
-            'hideOneItem' => false,
-            'options' => [
-                'class' => 'home-nav',
-            ],
-        ]); ?>
-        <?php
-    }
-?>
-    <?= Nav::widget([
-    'hideOneItem' => false,
-    'items' => [
-        [
-            'label' => Yii::t('anakin', 'ANAKIN_DASHBOARD_SKYPE'),
-            'url' => 'skype:danozzzz',
-            'linkOptions' => [
-                'id' => 'skype',
-            ],
-        ],
-    ],
-    'options' => [
-        'class' => 'home-nav',
-    ],
-]); ?>
+    <?= Nav::make()
+        ->class('home-nav')
+        ->addItem(NavItem::make()
+            ->label(Yii::t('anakin', 'ANAKIN_DASHBOARD_SKYPE'))
+            ->url('skype:danozzzz')); ?>
     <div class="home-footer">
         <p><?= Yii::t('anakin', 'ANAKIN_DASHBOARD_CONTACT'); ?></p>
         <p>