+++ src/Modules/Admin/Widgets/Navs/AnakinAsideMenu.php
@@ -6,10 +6,12 @@ namespace Hirtz\Anakin\Modules\Admin\Widgets\Navs;
 
 use Hirtz\Skeleton\Html\Aside;
 use Hirtz\Skeleton\Modules\Admin\Widgets\Navs\AsideMenu;
+use Override;
 use Stringable;
 
 class AnakinAsideMenu extends AsideMenu
 {
+    #[Override]
     protected function renderContent(): Stringable
     {
         return Aside::make()